agent/requirements.txt
# Core dependencies for the VM agent
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
psutil==5.9.6
python-dotenv==1.0.0

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# VM management
libvirt-python==10.0.0

# Logging and utilities
colorlog==6.7.0
schedule==1.2.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1